### 2017-01-11 #159 @whedon
[PRE REVIEW]: RiskPortfolios
**Submitting author:** @ArdiaD (<a href="http://orcid.org/0000-0003-2823-782X">David Ardia</a>)
**Repository:** <a href="https://github.com/ArdiaD/RiskPortfolios" target ="_blank">https://github.com/ArdiaD/RiskPortfolios</a>
**Version:** 2.00.11
