### 2017-01-11 c28604f
+ small fix in CONTRIBUTING.md
