+++ CONTRIBUTING.md
@@ -1,7 +1,7 @@
 Contribute
 ==========
 
-Contributions to `RiskPortfolios` are very welcome. Fork `RiskPortfolios` at [github](https://github.com/ArdiaD/RiskPortfolios), implement 
+Contributions to `RiskPortfolios` are very welcome. Fork `RiskPortfolios` at [GitHub](https://github.com/ArdiaD/RiskPortfolios), implement 
 your feature and send us
 a pull request. Also, please use the GitHub [issue tracker](https://github.com/ArdiaD/RiskPortfolios/issues)
 to let us know about bugs or feature requests, or if you have problems or questions regarding `RiskPortfolios`.
