### 2017-01-13 742657b
small fix in DESCRIPTION to match CRAN standards
