### 2017-01-20 98bae47
+ improved doc
