+++ R/optimalPortfolio.R
@@ -52,7 +52,7 @@
 #' \item \code{UB} lower boundary for the weights. Default: \code{UB = NULL}. 
 #' 
 #' \item \code{w0} starting value for the optimizer. Default: \code{w0 = NULL} takes the 
-#' equally-weighted portfolio as a starting value. When \code{LB} and \code{UB} are provided, it it set to 
+#' equally-weighted portfolio as a starting value. When \code{LB} and \code{UB} are provided, it is set to 
 #' mid-point of the bounds.
 #' 
 #' \item \code{gross.c} gross exposure constraint. Default: \code{gross.c = 1.6}. 
@@ -65,7 +65,7 @@
 #' @param Sigma a \eqn{(N \times N)}{(N x N)} covariance matrix.
 #' @param mu a \eqn{(N \times 1)}{(N x 1)} vector of expected returns. Default:
 #' \code{mu = NULL}.
-#' @param semiDev a vector \eqn{(N \times 1)}{(N x 1)} of semideviation.
+#' @param semiDev a vector \eqn{(N \times 1)}{(N x 1)} of semideviations.
 #' Default: \code{semiDev = NULL}.
 #' @param control control parameters (see *Details*).
 #' @return A \eqn{(N \times 1)}{(N x 1)} vector of optimal portfolio weights.
@@ -77,39 +77,48 @@
 #' 
 #' Ardia, D., Boudt, K. (2015). 
 #' Implied expected returns and the choice of a mean-variance efficient portfolio proxy.
-#' \emph{Journal of Portfolio Management} \bold{41}(4), pp.66--81. \doi{10.3905/jpm.2015.41.4.068}.
+#' \emph{Journal of Portfolio Management} \bold{41}(4), pp.66--81. 
+#' \doi{10.3905/jpm.2015.41.4.068}
 #' 
 #' Ardia, D., Bolliger, G., Boudt, K., Gagnon-Fleury, J.-P. (2016).  
-#' The Impact of covariance misspecification in risk-based portfolios.  
-#' \emph{Working paper}. URL \url{https://papers.ssrn.com/sol3/papers.cfm?abstract_id=2650644}
+#' \emph{The Impact of covariance misspecification in risk-based portfolios}.  
+#' Working paper. 
+#' \url{https://papers.ssrn.com/sol3/papers.cfm?abstract_id=2650644}
 #' 
 #' Choueifaty, Y., Coignard, Y. (2008).  
 #' Toward maximum diversification.
-#' \emph{Journal of Portfolio Management} \bold{35}(1), pp.40--51. \doi{10.3905/JPM.2008.35.1.40}.
+#' \emph{Journal of Portfolio Management} \bold{35}(1), pp.40--51. 
+#' \doi{10.3905/JPM.2008.35.1.40}
 #' 
 #' Choueifaty, Y., Froidure, T., Reynier, J. (2013).  
 #' Properties of the most diversified portfolio.  
-#' \emph{Journal of Investment Strategies} \bold{2}(2), pp.49--70. \doi{10.21314/JOIS.2013.033}.
+#' \emph{Journal of Investment Strategies} \bold{2}(2), pp.49--70. 
+#' \doi{10.21314/JOIS.2013.033}
 #' 
 #' Das, S., Markowitz, H., Scheid, J., Statman, M. (2010).  
 #' Portfolio optimization with mental accounts.  
-#' \emph{Journal of Financial and Quantitative Analysis} \bold{45}(2), pp.311--334. \doi{10.1017/S0022109010000141}.
+#' \emph{Journal of Financial and Quantitative Analysis} \bold{45}(2), pp.311--334. 
+#' \doi{10.1017/S0022109010000141}
 #' 
[…]
