+++ R/covEstimation.R
@@ -11,17 +11,17 @@
 #' \code{'diag'} and \code{'large'} where: 
 #' 
 #' \code{'naive'} is used to compute 
-#' the naive covariance matrix. 
+#' the naive (standard) covariance matrix. 
 #' 
 #' \code{'ewma'} is used to compute the exponential weighting moving average covariance matrix. The following formula is used 
-#' to compute de ewma covariance matrix:
+#' to compute the ewma covariance matrix:
 #' \deqn{\Sigma_t := \lambda \Sigma_{t-1} + (1-\lambda)r_{t-1}r_{t-1}}{Sigma[t]
 #' := lambda * Sigma[t-1] + (1-lambda) r[t-1]'r[t-1]}
 #' where \eqn{r_t} is the \eqn{(N \times 1)}{(N x 1)} vector of returns at time
-#' \eqn{t}. Note that the data must be sorted from the oldest to the latest.
+#' \eqn{t}. Note that the data must be sorted from the oldest to the latest. See RiskMetrics (1996)
 #' 
 #' \code{'factor'} is used to compute the covariance matrix estimation using a
-#' K-factor approach.
+#' K-factor approach. See Harman (1976).
 #' 
 #' \code{'lw'} is a weighted average of the sample covariance matrix and a
 #' 'prior' or 'shrinkage target'. The prior is given by a one-factor model and
@@ -30,7 +30,7 @@
 #' 
 #' \code{'const'} is a weighted average of the sample covariance matrix and a
 #' 'prior' or 'shrinkage target'.  The prior is given by constant correlation
-#' matrix. See Ledoit, O., Wolf, M. (2002).
+#' matrix. See Ledoit and Wolf (2002).
 #' 
 #' \code{'cor'} is a weighted average of the sample covariance matrix and a
 #' 'prior' or 'shrinkage target'.  The prior is given by the constant
@@ -39,11 +39,11 @@
 #' \code{'oneparm'} is a weighted average of the sample covariance matrix and a
 #' 'prior' or 'shrinkage target'.  The prior is given by the one-parameter
 #' matrix. All variances are the same and all covariances are zero. 
-#' See Ledoit, O., Wolf, M. (2004).
+#' See Ledoit and Wolf (2004).
 #' 
 #' \code{'diag'} is a weighted average of the sample covariance matrix and a
 #' 'prior' or 'shrinkage target'.  The prior is given by a diagonal matrix. 
-#' See Ledoit, O., Wolf, M. (2002).
+#' See Ledoit and Wolf (2002).
 #' 
 #' \code{'large'} This estimator is a weighted average of the sample covariance
 #' matrix and a 'prior' or 'shrinkage target'. Here, the prior is given by a
@@ -54,7 +54,7 @@
 #' infinity, but Monte-Carlo simulations show that it works well for values as
 #' low as 10. The main advantage is that this estimator is guaranteed to be
 #' invertible and well-conditioned even if variables outnumber observations. 
-#' See Ledoit, O., Wolf, M. (2004).
+#' See Ledoit and Wolf (2004).
 #' 
 #' \code{'bs'} is the Bayes-Stein estimator for the covariance matrix given by
 #' Jorion (1986).
@@ -72,25 +72,34 @@
 #' @note Part of the code is adapted from the Matlab code by Ledoit and Wolf (2014).
 #' @author David Ardia, Kris Boudt and Jean-Philippe Gagnon Fleury.
 #' @references 
[…]
