### 2017-01-20 117affc
+ fix CRAN README
