+++ README CRAN.md
@@ -1,0 +1,4 @@
+`RiskPortfolios` is an R package for constructing risk-based portfolios dedicated to portfolio managers 
+and quantitative analysts. It provides a set of functionalities to build mean-variance, minimum variance, inverse-volatility weighted, 
+equal-risk-contribution, maximum diversification, and risk-efficient portfolios. As risk-based portfolios are
+mainly based on covariances, the package also provides a large set of covariance matrix estimators.
