### 2017-01-20 #159 @ArdiaD
@arfon Also, MatthieuStigler on GitHub would be able to review, if he has enough time at disposal.
