### 2017-01-21 #159 @arfon
@enricoschumann @MatthieuStigler would either of you be willing to review this submission for JOSS? You can see an overview of our review process here: http://joss.theoj.org/about#reviewer_guidelines
