### 2017-01-26 #159 @joshuaulrich
Thank you very much for the consideration, but I have to decline because I will be very short on time for the next few weeks.  I'm working to finish a project with a deadline looming in mid-February.  Please do consider me for other finance-related submissions though.
