### 2017-01-26 #159 @ArdiaD
Thanks @MatthieuStigler and @joshuaulrich for your feedback. @arfon what should we do then? Do you want me to try to provide additional references?
