### 2017-01-27 #159 @arfon
> Thanks @MatthieuStigler and @joshuaulrich for your feedback. @arfon what should we do then? Do you want me to try to provide additional references?

Yes please 🙏
