### 2017-01-28 #159 @arfon
Thanks @ArdiaD.

> Ok, then I can imagine @braverock or @peterccarl, or @bpfaff. Also, @enricoschumann is OK to review. I know all of them, and I am convinced they will make the right comments to improve the package.

👋  all. Would any of you be willing to review this submission for JOSS? You can read more about the reviewer process here: http://joss.theoj.org/about#reviewer_guidelines
