### 2017-01-28 #159 @braverock
@arfon, @ArdiaD I'd be happy to take a look at the submission.

It isn't clear to me from your reviewer guidelines where my feedback should go.
