### 2017-01-28 #159 @arfon
> It isn't clear to me from your reviewer guidelines where my feedback should go.

Many thanks for agreeing to review this submission @braverock! This (pre review) issue is used to find a reviewer, the actual review thread with the checklist for you to work through is in #171.

Please let me know if you have any questions as you make progress.
