### 2017-01-28 #171 @braverock
Current repository version is 2.00.12, not the 2.00.11 in the checklist.
