### 2017-01-28 8ed4e54
+ version updated for JOSS
