### 2017-02-02 #171 @arfon
> @arfon, I have completed my review of the package code. Everything appears to be in order. In addition to providing some functionality that exists in other packages, the authors have added several more cases which were not well-represented in existing packages for the purposes of creating risk constrained portfolios.

Excellent. Thanks @braverock :zap:
