### 2017-02-03 #171 @arfon
Many thanks for your review @braverock ✨

@ArdiaD your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00171 ⚡️ 🚀 💥
